+++ scripts/dictionaries.sh
@@ -1,11 +1,11 @@
 #!/usr/bin/env bash
 pushd .
-mkdir -p $(pwd)/server/dictionaries
-cd $(pwd)/server/dictionaries
+mkdir -p $(pwd)/client/resources
+cd $(pwd)/client/resources
 mkdir -p temp
 pushd temp
 curl -O ftp://ftp.iucr.org/pub/cifdics/cifdic.register
-node -e 'require("../../out/dictionaries").readRegister()' > temp.sh
+node -e 'require("../../../server/out/dictionaries").readRegister()' > temp.sh
 popd
 sh ./temp/temp.sh
 popd