+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.7.0 - ...
+- Add commands for adding, removing and listing CIF dictionaries
 - Allow double quotes inside double-quoted strings
 - Add developer guide
 - Revalidate documents on configuration change